+++ app/Http/Controllers/Product/ItemsController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Item;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class ItemsController extends Controller
 {
@@ -21,16 +22,25 @@ class ItemsController extends Controller
 
     public function store(Request $request)
     {
+        if (!Auth::user()->is_admin) {
+            return response()->json(['error' => 'Unauthorized. Only admin can create items.'], 403);
+        }
         return Item::create($request->all());
     }
 
     public function update(Request $request, Item $item)
     {
+        if (!Auth::user()->is_admin) {
+            return response()->json(['error' => 'Unauthorized. Only admin can update items.'], 403);
+        }
         return $item->update($request->all());
     }
 
     public function destroy(Item $item)
     {
+        if (!Auth::user()->is_admin) {
+            return response()->json(['error' => 'Unauthorized. Only admin can delete items.'], 403);
+        }
         return $item->delete();
     }
 }
